make build crossplatform
This commit streamlines the build process using the mod build config package, which automates references to Stardew Valley and SMAPI depending on the detected install path and OS. That lets other developers compile the project on any computer/OS without changing hardcoded paths in the project file. This commit also replaces one of the PowerShell scripts with a crossplatform MSBuild task, and disables the other when the OS isn't Windows.